requirements.txt
spotipy~=2.16.1
pandas~=1.2.0
numpy~=1.19.4
tensorflow~=2.14.0
opencv-python~=4.4.0.46
scikit-learn~=0.24.0
Pillow~=8.1.0